# business-core/docs/api/business-core-openapi.yaml
openapi: 3.1.0
info:
  title: Business Core API
  version: "1.0.0"
  description: |
    API du Business Core — noyau métier générique intégré au kernel RT-Comops.

    Le Business Core expose un modèle de métier piloté par les métadonnées :
    le développeur déclare des Types Métier (le modèle), crée des Entreprises
    (les instances), et déclenche des Opérations. Le Business Core orchestre
    le kernel en façade.

    ## Authentification
    Chaque requête porte la clé Business Core du développeur via les en-têtes
    `X-BC-Client-Id` et `X-BC-Api-Key`. Le secret kernel n'est jamais exposé.

    ## Conventions
    - Versionnement de l'API dans l'URL (`/v1`), distinct de la version d'un Type Métier.
    - Opérations synchrones par défaut (`200`), différées en `202` avec une trace de suivi.
    - Erreurs au format RFC 7807 (Problem Details), enrichies de champs métier.
  contact:
    name: Équipe Business Core
  license:
    name: Usage académique
servers:
  - url: https://business-core.example.com
    description: Serveur du Business Core
security:
  - BusinessCoreKey: []
...
paths:
...
  /v1/business-types/{typeId}/versions/{versionNumber}/rules:
    parameters:
      - $ref: '#/components/parameters/TypeId'
      - $ref: '#/components/parameters/VersionNumber'
    post:
      tags: [Contenu de version]
      summary: Déclarer une règle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRule'
      responses:
        '201':
          description: Règle déclarée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
    get:
      tags: [Contenu de version]
      summary: Lister les règles de Type
      responses:
        '200':
          description: Liste des règles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rule'
  /v1/business-types/{typeId}/versions/{versionNumber}/rules/{ruleId}:
    parameters:
      - $ref: '#/components/parameters/TypeId'
      - $ref: '#/components/parameters/VersionNumber'
      - name: ruleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Contenu de version]
      summary: Consulter une règle de Type
      responses:
        '200':
          description: La règle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          $ref: '#/components/responses/Problem'
    put:
      tags: [Contenu de version]
      summary: Modifier une règle de Type
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRule'
      responses:
        '200':
          description: Règle mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Contenu de version]
      summary: Supprimer une règle de Type
      responses:
        '204':
          description: Règle supprimée
        '404':
          $ref: '#/components/responses/Problem'
...
  /v1/businesses/{businessId}/rules:
    parameters:
      - $ref: '#/components/parameters/BusinessId'
    post:
      tags: [Entreprises]
      summary: Ajouter une règle locale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRule'
      responses:
        '201':
          description: Règle locale ajoutée
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
    get:
      tags: [Entreprises]
      summary: Lister les règles locales
      responses:
        '200':
          description: Liste des règles locales
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Rule'
  /v1/businesses/{businessId}/rules/{ruleId}:
    parameters:
      - $ref: '#/components/parameters/BusinessId'
      - name: ruleId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      tags: [Entreprises]
      summary: Consulter une règle locale
      responses:
        '200':
          description: La règle
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          $ref: '#/components/responses/Problem'
    put:
      tags: [Entreprises]
      summary: Modifier une règle locale
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRule'
      responses:
        '200':
          description: Règle mise à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Rule'
        '404':
          $ref: '#/components/responses/Problem'
    delete:
      tags: [Entreprises]
      summary: Supprimer une règle locale
      responses:
        '204':
          description: Règle supprimée
        '404':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    BusinessCoreKey:
      type: apiKey
      in: header
      name: X-BC-Api-Key
      description: Clé d'API du développeur. Accompagnée de l'en-tête X-BC-Client-Id.
  parameters:
    TypeId:
      name: typeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    VersionNumber:
      name: versionNumber
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    BusinessId:
      name: businessId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Problem:
      description: Erreur au format RFC 7807
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    Problem:
      type: object
      description: Détail d'erreur RFC 7807, enrichi de champs métier.
      properties:
        type:
          type: string
          example: /errors/regle-violee
        title:
          type: string
          example: Règle métier non respectée
        status:
          type: integer
          example: 422
        detail:
          type: string
          example: Un document ordonnance est requis pour cette vente.
        instance:
          type: string
          example: /v1/businesses/abc/operations/vente:execute
        violatedRule:
          type: string
          description: Code de la règle violée (extension métier)
        requiredAction:
          type: string
          enum: [BLOQUER, EXIGER, VALIDER, AJUSTER, ALERTER, DEROGER]
        requiredDocument:
          type: string
...
    CreateRule:
      type: object
      required: [declencheur, effet]
      properties:
        declencheur:
          type: string
          example: AVANT_VENTE
        condition:
          type: string
          description: Condition déclarative (niveau 1 paramétré)
          example: "offre.categorie == 'ordonnance'"
        effet:
          type: string
          enum: [BLOQUER, EXIGER, VALIDER, AJUSTER, ALERTER, DEROGER]
        rolesAutorisesADeroger:
          type: array
          items:
            type: string
    Rule:
      allOf:
        - $ref: '#/components/schemas/CreateRule'
        - type: object
          properties:
            id:
              type: string
              format: uuid
            portee:
              type: string
              enum: [TYPE, ENTREPRISE]
...